requirements.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
distribute==0.6.24
pycrypto==2.6
wsgiref==0.1.2
